python/agents/youtube-analyst/pyproject.toml
[project]
name = "youtube-analyst"
version = "0.1.0"
description = "A Gemini-powered agent for analyzing YouTube data."
authors = [
    { name = "Pili Hu", email = "hupili@google.com" },
    { name = "Jasmine Tong", email = "jasminetong@google.com" },
    { name = "Kun Wang", email = "kunwk@google.com" },
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=0.1.0",
    "google-genai>=0.2.0",
    "google-api-python-client",
    "matplotlib",
    "plotly",
    "pandas",
    "pydantic",
    "pydantic-settings",
]

[dependency-groups]
dev = [
    "pytest",
]
